+++ src/lib/feature/publicPage/ui/PublicPosts/PublicPosts.tsx
@@ -4,7 +4,7 @@ import { Typography } from '@/components/ui/typography/Typography'
 import Image from 'next/image'
 import Link from 'next/link'
 import { formatTimeAgo } from '@/utils/formatTimeAgo'
-import ExpandableText from '@/components/ui/expandableText/ExpandableText'
+import ExpandableText from '@/lib/feature/publicPage/ui/expandableText/ExpandableText'
 import { POSTS_ON_MAIN_PAGE } from '@/constants'
 import { useGetPublicPostsQuery } from '@/lib/feature/posts/api/postsApi'
 import { UserPostsResponse } from '@/lib/feature/posts/api/postsApi.types'
